add update capability and statistics

+++ mysite/pics/tasks.py
@@ -6,7 +6,7 @@ import json
 import os
 import time
 from celery import shared_task
-from pics.models import Photo
+from pics.models import Photo, Statistics
 from pics.settings import (VIEW_THRESHOLD, DOWNLOAD_PATH, FONT_PATH,
                            TEST_LIMIT, FLICKR_USER_ID)
 from pics.flickr_utils import (flickr_connect, get_flickr_photo, 
@@ -95,51 +95,7 @@ def update_one_record(photo):
 
     photo.save()
 
-def create_photo_object(flickr, photo_id, info_dict):
-    """Create a Photo object and save it using the relevent
-    metadata for the photo from flickr.
-    """
-    photo = Photo()
-    photo.pic_id = photo_id
-    title = info_dict['photo']['title']['_content']
-    # quotes in the title confuse things
-    title = title.replace("'", '')
-    title = title.replace('"', '')
-    photo.title = title
-    photo.date_taken = '%s+00' % info_dict['photo']['dates']['taken']
-    photo.date_posted = time.strftime('%Y-%m-%d %H:%M:%S+00', time.localtime(
-                            int(info_dict['photo']['dates']['posted'])))
-    photo.date_updated = time.strftime('%Y-%m-%d %H:%M:%S+00', time.localtime(
-                             int(info_dict['photo']['dates']['lastupdate'])))
-    photo.view_count = int(info_dict['photo']['views'])
-    tags = info_dict['photo']['tags']
-    for tag in tags['tag']:
-        if tag['raw'] == 'wallpaper':
-            photo.wallpaper = True
-            break
-
-    retries = 5
-    photo.source_url = None
-    while retries > 0:
-        try:
-            sizes = flickr.photos.getSizes(photo_id=photo_id,
-                                           format='json')
-            retries = 0
-            sizes_dict = json.loads(sizes.decode("utf-8"))
-            source = None
-            for entry in sizes_dict['sizes']['size']:
-                if entry['label'] == 'Original':
-                    photo.source_url = entry['source']
-                    break
-        except flickrapi.exceptions.FlickrError:
-            retries -= 1
-            time.sleep(0.1)
-    
-    if photo.source_url:
-        photo.save()
-
-@shared_task(bind=True)
-def update_pics_db(self):
+def update_db_from_flickr(self, rebuild):
     """Load the database with information on all of the pictures
     in flickr.
     """
@@ -153,8 +109,11 @@ def update_pics_db(self):
         total = get_public_count(flickr)
     progress = 0
 
-    # delete the info currently in the table
-    Photo.objects.all().delete()
+    # on a rebuild, delete the info currently in the table
+    if rebuild is True:
+        Photo.objects.all().delete()
+
+    edit_count = 0
 
     for photo in flickr.walk(user_id=FLICKR_USER_ID):
         photo_id = photo.get('id')
@@ -163,7 +122,9 @@ def update_pics_db(self):
             try:
                 info_dict = get_flickr_photo(flickr, photo_id)
                 if info_dict['photo']['visibility']['ispublic'] == 1:
-                    create_photo_object(flickr, photo_id, info_dict)
+                    status = Photo.create_or_update(flickr, photo_id, info_dict)
+                    if status == 'edit':
+                        edit_count += 1
                 retries = 0
             except flickrapi.exceptions.FlickrError:
                 retries -= 1
@@ -171,10 +132,29 @@ def update_pics_db(self):
         progress_recorder.set_progress(progress, total)
         if progress < total:
             progress += 1
-        if TEST_LIMIT > 0 and progress > TEST_LIMIT:
+        if TEST_LIMIT > 0 and progress >= TEST_LIMIT:
             break
     progress_recorder.set_progress(total, total)
+    Statistics.create(edit_count)
+    return
+
+@shared_task(bind=True)
+def update_pics_db(self):
+    """Load the database with information on all of the pictures
+    in flickr.
+    """
+    update_db_from_flickr(self, rebuild=False)
     return True
+
+
+@shared_task(bind=True)
+def rebuild_pics_db(self):
+    """Load the database with information on all of the pictures
+    in flickr.
+    """
+    update_db_from_flickr(self, rebuild=True)
+    return True
+
 
 @shared_task(bind=True)
 def download_pics_task(self):
